requirements.txt
asana==3.2.1
twilio==8.10.0
python-dotenv==1.0.0
flask==3.0.0
schedule==1.2.0
requests==2.31.0
pydantic==2.5.0
loguru==0.7.2 